New option to trigger alert on failed restarts
Current versions of puppet treat resources that failed to restart as different
kind of failures and they are not necessarily reported, there is puppet bug
PUP-2280 that explains it in more detail. Until it is fixed in version 5 of
puppet it would be useful to see such failures in monitoring, so I added option
to report this failure. Option is by default off to not change usual behavior.